fix: change keyring `ServiceName` to camel case
This patch changes the `Freedesktop.SecretService` service name to be 
compatible with `Freedesktop D-bus` specification.

See: https://dbus.freedesktop.org/doc/dbus-specification.html#message-protocol-names

When using `Freedesktop.SecretService` as secret backend, the collection
created will be the same as the service name, and character `-` will be
converted to `_2d` when caching the token, however when retrieving the
token the collection name expected will still be `k8s-pixy-auth`.
Therefore `k8s-pixy-auth` will not find the cached token and start the
authentication journey again.

Signed-off-by: Andrew Zheng <andrew.zheng@solibri.com>